Add starting GPIO setup commands
To setup a GPIO pin, the struct "gpio_config_t" is needed. This struct
contains various different values such as "intr_type" and "mode". For
the purposes of this project, I think the only values needed would be:
- Mode
- Pin Bit Mask
- Pull Down En
- Pull Up En

I am still not sure what pull down/up does, but I can mess with that
later.